+++ app/mailers/application_mailer.rb
@@ -1,4 +1,6 @@
 class ApplicationMailer < ActionMailer::Base
-  default from: "from@example.com"
+  default from: "contato@mudamos.org",
+          reply_to: "contato@mudamos.org"
+
   layout 'mailer'
 end